Add : Add exception for deleteById Menu

+++ src/main/java/com/gotcha/earlytable/domain/menu/MenuService.java
@@ -93,6 +93,10 @@ public class MenuService {
      * 메뉴 삭제 서비스 메서드
      */
     public void deleteMenu(Long menuId) {
+        if(!menuRepository.existsById(menuId)){
+            throw new NotFoundException(ErrorCode.NOT_FOUND);
+        }
+        
         menuRepository.deleteById(menuId);
     }
 }
